chore: compare repro with node
