+++ examples/repro/README.md
@@ -1,10 +1,12 @@
 # repro
 
 ```sh
-node test.js
-[debug:import] file:///home/hiroshi/code/personal/vite-environment-examples/examples/repro/src/thing.ts
-[mod1.thing] { k: 'v' }
-[debug:import] file:///home/hiroshi/code/personal/vite-environment-examples/examples/repro/src/thing.ts
-[mod2.thing] { k: 'v' }
-[mod1.thing === mod2.thing] false
+$ node test-vite.js
+[debug:import] file:///home/hiroshi/code/personal/vite-environment-examples/examples/repro/src/thing.js
+[debug:import] file:///home/hiroshi/code/personal/vite-environment-examples/examples/repro/src/thing.js
+false
+
+$ node test-node.js
+[debug:import] file:///home/hiroshi/code/personal/vite-environment-examples/examples/repro/src/thing.js
+true
 ```
