+++ examples/repro/src/entry.js
@@ -1,0 +1,12 @@
+import { thing } from "./thing.js";
+
+export function getThing() {
+  return thing;
+}
+
+/**
+ * @param {string} id
+ */
+export function myImport(id) {
+  return import(/* @vite-ignore */ id);
+}
