--- examples/repro/src/entry.ts
@@ -1,3 +1,0 @@
-export function myImport(id: string) {
-  return import(/* @vite-ignore */ id);
-}
